requirements/requirements.txt
absl-py>=1.4.0
chex>=0.1.82
dm-env>=1.6
dm-haiku>=0.0.10
dm-tree>=0.1.8
jax>=0.4.14
jaxlib>=0.1.14
numpy>=1.25.0
optax>=0.1.7
